Enable and fix RSpec/DescribedClassModuleWrapping
